"Revert simesannpcbant.tra back to English
The reason is this is a crossmod for an old BGT Sime version who was a ranger.
This mod is probably never going to exist for EET.
I don't want to translate the content that probably never going to be used,
but I don't want to have files with "НП: " to confuse myself.
Thus, I decided to revert this file in English.

As a general though - if something is keept in English, it is probably
should not be removed for some reasons.